Leetcode 215: Kth Largest Element Of An Array
Approach 3: Counting Sort
Time: O(n)
Space: O(n)

It employs counting sort to count the frequency of each element in the array.
The algorithm then iterates through the counting array from the highest index to the lowest,
decrementing k until it becomes less than or equal to 0, and returns the corresponding value in the original array.